Backend/Backend-Users/requirements.txt
anyio==3.7.0
bcrypt==4.0.1
black==23.3.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
dnspython==2.3.0
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.98.0
h11==0.14.0
httptools==0.5.0
idna==3.4
motor==3.1.1
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
pip-review==1.3.0
platformdirs==3.8.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
